requirements.txt
pillow~=9.1.0
numpy~=1.22.3
ray~=1.13.0
fastapi~=0.77.1
setuptools~=62.2.0
gym~=0.21.0
wandb~=0.12.11
torchinfo~=1.6.3
pysimplegui~=4.60.0
igraph~=0.9.10
matplotlib~=3.5.2
pandas~=1.4.2
jpype1>=1.3.0
igraph>=0.9.10
tqdm>=4.64.0
ipython>=8.3.0
torch>=1.11.0